Switch from curl to wget
The latest version of `appropriate/curl` does not handle the
`--retry-connrefused` argument correctly, and older versions are
unavailable.  Switching to `wget` was the quickest fix, and I've uploaded
the image we're using to our own repository on Quay so we don't run into
this issue again even if `wget` changes.